Major code restructuring in src and tests
Most importandly:
  - Vectors are not matrices any more, but are separate objects.
  - Both matrices and vectors are derived off Indexable_i.
  - New implementation of ArmadilloVector.
  - Numeric data generation now part of the linalgwrap library under
    TestingUtils.hh

+++ tests/RangeTests.cc
@@ -38,7 +38,8 @@ struct RangeTests {
 
         if (is_valid) {
             range_type r{t1, t2};
-            RC_ASSERT(r.length() == static_cast<size_type>(t2 - t1));
+            RC_ASSERT(r.length() ==
+                      static_cast<size_type>(static_cast<ptrdiff_t>(t2) - t1));
         } else {
 #ifdef DEBUG
             RC_ASSERT_THROWS_AS((Range<T>{t1, t2}), krims::ExcTooLarge<T>);
